==> notebooks/02_gemma_responses.ipynb <==
# Prompt Gemma Model

Prompting Gemma and Experimentation

Load Models and necessary libraries

import pandas as pd
import transformers
import inseq
import torch
from pathlib import Path

torch.cuda.is_available()

from huggingface_hub import login
login(token="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")

import torch
from transformers import AutoTokenizer, AutoModelForCausalLM

MODEL_ID = "google/gemma-2b-it"

tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
model = AutoModelForCausalLM.from_pretrained(MODEL_ID, device_map="auto")

aj = Path("stand_up_dataset/aj/AJ_TP_text_15.txt").read_text()
aj2 = Path("stand_up_dataset/aj2/AJ_TP_text_25.txt").read_text()
aw = Path("stand_up_dataset/aw/AW_BC_text_04.txt").read_text()
aw2 = Path("stand_up_dataset/aw2/AW_BC_text_25.txt").read_text()
cp = Path("stand_up_dataset/cp/CP_OTG_text_18.txt").read_text()
cp2 = Path("stand_up_dataset/cp2/CP_OTG_text_07.txt").read_text()
dg = Path("stand_up_dataset/dg/DG_W_text_11.txt").read_text()
dg2 = Path("stand_up_dataset/dg2/DG_W_text_07.txt").read_text()
hm = Path("stand_up_dataset/hm/HM_HK_text_05.txt").read_text()
hm2 = Path("stand_up_dataset/hm2/HM_HK_text_22.txt").read_text()
iss = Path("stand_up_dataset/is/IS_EM_text_18.txt").read_text()
is2 = Path("stand_up_dataset/is2/IS_EM_text_20.txt").read_text()
jg = Path("stand_up_dataset/jg/JG_QT_text_03.txt").read_text()
jg2 = Path("stand_up_dataset/jg2/JG_QT_text_14.txt").read_text()
ji = Path("stand_up_dataset/ji/JL_IHM_text_11.txt").read_text()
ji2 = Path("stand_up_dataset/ji2/JL_IHM_text_02.txt").read_text()
jm = Path("stand_up_dataset/jm/JB_KGRC_text_12.txt").read_text()
jm2 = Path("stand_up_dataset/jm2/JB_KGRC_text_06.txt").read_text()
jy = Path("stand_up_dataset/jy/JY_GD_text_09.txt").read_text()
jy2 = Path("stand_up_dataset/jy2/JY_GD_text_19.txt").read_text()
lck = Path("stand_up_dataset/lck/LCK_SY_text_06.txt").read_text()
lck2 = Path("stand_up_dataset/lck2/LCK_SY_text_20.txt").read_text()
nb = Path("stand_up_dataset/nb/NB_TGAA_text_09.txt").read_text()
nb2 = Path("stand_up_dataset/nb2/NB_TGAA_text_14.txt").read_text()
nb_tk = Path("stand_up_dataset/nb_tk/NB_TK_text_20.txt").read_text()
nb_tk2 = Path("stand_up_dataset/nb_tk2/NB_TK_text_17.txt").read_text()
rp = Path("stand_up_dataset/rp/RP_DP_text_06.txt").read_text()
rp2 = Path("stand_up_dataset/rp2/RP_DP_text_21.txt").read_text()
sm = Path("stand_up_dataset/sm/SM_IGT_text_04.txt").read_text()
sm2 = Path("stand_up_dataset/sm2/SM_IGT_text_13.txt").read_text()
tn = Path("stand_up_dataset/tn/TN_AD_text_16.txt").read_text()
tn2 = Path("stand_up_dataset/tn2/TN_AD_text_14.txt").read_text()
ts = Path("stand_up_dataset/ts/TS_D_text_09.txt").read_text()
ts2 = Path("stand_up_dataset/ts2/TS_D_text_15.txt").read_text()

transcripts = [aj, aj2, aw, aw2, cp, cp2, dg, dg2, hm, hm2, iss, is2, jg, jg2, ji, ji2, jm, jm2, jy, jy2, lck, lck2, nb, nb2, nb_tk, nb_tk2, rp, rp2, sm, sm2, tn, tn2, ts, ts2]

comedian_names = ["Anthony_Jeselnik", 
                  "Anthony_Jeselnik_2", 
                  "Ali_Wong", 
                  "Ali_Wong_2",
                  "Chelsea_Peretti",
                  "Chelsea_Peretti_2",
                  "Donald_Glover",
                  "Donald_Glover_2",
                  "Hasan_Minhaj",
                  "Hasan_Minhaj_2",
                  "Iliza_Shlesinger",
                  "Iliza_Shlesinger_2",
                  "Jim_Gaffigan",
                  "Jim_Gaffigan_2",
                  "Joe_List",
                  "Joe_List_2",
                  "John_Mulaney",
                  "John_Mulaney_2",
                  "Jimmy_Yang",
                  "Jimmy_Yang_2",
                  "Louis_CK",
                  "Louis_CK_2",
                  "Nate_Bargatze",
                  "Nate_Bargatze_2",
                  "Nate_Bargatze_TK",
                  "Nate_Bargatze_TK_2",
                  "Russell_Peters",
                  "Russell_Peters_2",
                  "Sam_Morril",
                  "Sam_Morril_2",
                  "Trevor_Noah",
                  "Trevor_Noah_2",
                  "Tom_Segura",
                  "Tom_Segura_2"]

instruction = 

results_dict = {}
for comedian, transcript in zip(comedian_names, transcripts):
    prompt = f"'''{transcript}'''\n\n{instruction}"
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=120)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(prompt, "").strip()
    results_dict[comedian] = response

pandas_dict = {"Anthony_Jeselnik": 
               ["When I was a kid, I used to fantasize about getting older, growing up and having money, and buying my mom nice things for her birthday.",
                "My mom actually should've been on one of the planes that crashed on 9/11.",
                "When I was a kid, like nine years old, I'd come home after school. Once in a while, I'd bring a friend over to play with me at my house. Once in a while, I'd bring a black friend over. And when I did that,"],
                "Anthony_Jeselnik_2":
                ["I've never talked to a group of people without getting paid a lot of money. How should I handle this?",
                "And I know my grandma loved it too, because it combined her two favorite things: spending time with her grandchildren, and using the 'N' word.",
                "I'm a real comedian. I am a pure comedian. I think I'm one of the best comedians of all time."],
                "Ali_Wong": 
                ["It was like the worst experience of my life. It was so emotional.",
                 "Those of you under 25 probably don't know what that calculator is.",
                 "It was like the Tesla of my time.",
                 "I'm not trying to be a procrastinator anymore. Because according to Deepak-Oprah, that's not the way for me to achieve my optimum level of success."],
                "Ali_Wong_2":
                ["Life is not Rush Hour, the movie, OK?", 
                 "I'm just gonna recycle… go bald… go to the park, do this shit.",
                 "Old Asian ladies and the sun are like the Tupac and Biggie of longevity."],
                 "Chelsea_Peretti":
                 ["To the future!", "Uh-oh, it's us again.",
                  "My hubby made breakfast. My hubby fixed the door. My hubby is sleeping. He breathed in, he breathed out. He breathed in, he breathed out."],
                  "Chelsea_Peretti_2":
                  ["People say I look just like you.",
                   "It's actually to stop them from sweating, because why would you, you know, as a human?",
                   "I'm so drawn to you, but I can't get near you! This is a whole new kind of juicy booty.",
                   "Whoa, whoa, whoa!"],
                  "Donald_Glover":
                  ["I was like, what the fuck is wrong with this kid? What's going on with this kid?",
                   "The sweetest thing he was allowed was mints. He was just allowed to have mints.",
                   "It was crazy. I would take him to the park, right? Washington Square Park, and all the babysitters in New York for some reason are Trinidadian. They're all Trinidadian babysitters."],
                  "Donald_Glover_2":
                  ["Yeah, well, 'if anybody ever tries to rape me, I'm just gonna shit on 'em.' 'Yes, I will. I have no problem.'",
                   "I can poop whenever I want. Like, you can just poop whenever you want? Like, you got more superpowers than Shaft if you can poop whenever you want.",
                   "I feel like, me, I'm always like, come on, let's… hurry up, let's do this. And my butt's always like, stop rushing me, when"],
                   "Hasan_Minhaj":
                   ["Ah, Biju got slapped on his birthday!",
                    "That kid won't choke on camera. He's been slapped on camera.",
                    "Look at this kid's parents. Your son just won the Scripps National Spelling Bee. Look at his brother. His brother is like, 'I'm fucked. I'm fucked. The bar is way too high. I should kill myself.'",
                    "That's where he comes from. That is an Indian sociopath."],
                  "Hasan_Minhaj_2":
                  ["Alright, you guys are all killing it academically but I want you to know there is more to life than just getting into UC Berkeley.",
                   "I know, getting into Stanford.",
                   "No, you have to live a life worth talking about, which is why I'm making it mandatory for everyone in this class to go to prom.",
                   "Whatever, it's not going to happen. He can't do this."],
                  "Iliza_Shlesinger":
                  ["The very beginning of a new relationship is the most exciting part of the relationship.",
                   "It's the most nerve-wracking part of a new relationship when you're younger, is the first time a boy comes over to your apartment.",
                   "You want him to come in and be, like, “It's so homey. I'd like to stay forever.” Yes, come closer. Like, that's what you want."],
                  "Iliza_Shlesinger_2": 
                  ["You don't want to wear… what you wore during the day. Don't want work clothes. You don't want to wear your daytime clothes, 'cause… 'cause it's nighttime.",
                   "You don't want to wear your civilian clothes, okay. 'Cause you had a whole day. Maybe you sweat in them, they're gross."
                   "No means no. Kindergarteners get it."],
                  "Jim_Gaffigan":
                  ["That doesn't seem fair. I am now at the size, when I go in clothing stores sales people look at me like, “We got nothing for you. And you can't use our bathroom.”",
                    "Aw. Look at you try.",
                    "I like to sleep. It doesn't mean I wanna build a bed.",
                    "The truth is, I don't like to do anything."],
                  "Jim_Gaffigan_2":
                  ["That's right, I went, I went. It was in my apartment building and I needed the material.",
                   "Dog didn't know it was his birthday. The dog didn't know it had a birthday. The dog wasn't sure why people were in the apartment.",
                   "Someone made that up and we just went along with it."],
                   "Joe_List":
                   ["I was hoping it was three doctors saving money, by sharing a small office.",
                    "Oh, I guess it's pronounced tin-uh-dus.",
                    "It's pronounced t-ah-n-i-tus.",
                    "I'm hearing shit that's not even in the room."],
                    "Joe_List_2":
                    ["You can be in the second to last row. This one guy behind, you're like what a fucking idiot back there.",
                     "I wanted him to die, I swear to God.",
                     "It's like, if you were hungry on a plane, you were like, I'm hungry. Well you're all right? Yeah, yeah I'm hungry. I like to let people know when I'm hungry.",
                     "You're gonna die!"],
                    "John_Mulaney": 
                     ["It's good to give to charities, you know.",
                      "I took that stuff to Goodwill, right?",
                      "This one's shirts. I got a bunch of shirts! Take 'em away!",
                      "Well, if it's too big you can just wear it as a sleep shirt.",
                      "Oh, I get it. So rather than violate these meaningless politeness rules, I'll just go to bed in a smock like goddamn Ebenezer Scrooge."],
                    "John_Mulaney_2":
                    ["The greatest assembly of them all, once a year, Stranger Danger.",
                     "Detective JJ Bittenbinder's name is ridiculous.",
                     "He was a man most acquainted with misery.",
                     "He could look at a child and guess the price of their coffin."],
                    "Jimmy_Yang":
                    ["Asian people, we don't need Tinder anymore. We just go to BTS concerts. That's how we do that parking lot pimping.",
                     "I'm 5′ 5″. I just go to concerts to smell other people's armpits."],
                    "Jimmy_Yang_2":[
                      "It's really hard for me to watch TV with my dad, because he's trying to make me explain everything to him.",
                      "Old Asian people, they don't watch TV. They judge the TV.",
                      "He's just sitting there, arms folded, judging the TV like-He's made some random noises around the house.",
                      "It's just like, achoo! Oh! Ay, shit.Oh! I'm like, what the fuck, Dad, just have an orgasm?"],
                    "Louis_CK":
                    ["He was?",
                     "I don't think so.",
                     "Dude, Jesus couldn't be Jewish. Think about it.",
                     "That finally makes sense. I didn't even know.",
                     "Oh, fuck, that's me! I'm the little guy on it!"],
                  "Louis_CK_2":
                  ["It's sad how dumb they are.",
                   "They don't even know their own lives, they don't even — they can't even handle their own lives mentally.",
                   "I don't know anything now. Please help.",
                   "There it is, right there. It's right there."],
                  "Nate_Bargatze":
                  ["Did you name her after Harper Lee, the author of To Kill a Mockingbird?",
                   "I love how innocent it is. They cry over a tag in their shirt. I mean, they bawl. They don't like… It feels weird.",
                   "She just sits there on her iPad a lot. You know, that's the hard part. You got to get these kids off… You don't want technology all the time."],
                  "Nate_Bargatze_2":
                  ["Okay, learning it earlier than we used to, huh?",
                   "It's a whole new… I mean, it's unbelievable. They bring it home, you gotta watch a 40-minute YouTube video on Common Core math.",
                   "Oh! Just do that at the top. I don't even know what we're doing.",
                   "It's a long way to get at the same answer."],
                  "Nate_Bargatze_TK":
                  ["It goes clown, then magic.",
                    "The Easter Bunny was in the passenger seat.",
                    "Like, he couldn't sit normal, so his head was bent to the side.",
                    "Uh… I like to think about all the other people that saw that."],
                  "Nate_Bargatze_TK_2":
                  ["We gotta be there Monday for this wedding? It's a week?",
                   "Let's do it without it.",
                   "I didn't… How's your back worse than your front?",
                   "Yeah, that's my fault. I didn't… How's your back worse than your front?",
                   "I'm wildly overdressed. I look like I work there."],
                  "Russell_Peters":
                  ["That's really fucking racist, but since you asked, I have acid reflux.",
                   "There's no way you can be Indian and not have fucking acid reflux.",
                   "I'm looking at all of you, especially all the pudgy guys. You know exactly who has it.",
                   "You got it, don't you, yellow guy?"],
                  "Russell_Peters_2":
                  ["Identical twins are the only people that should be twins.",
                   "If I have twin girls, these are going to be my daughters, it's going to be Kate and Duplicate.",
                   "These are my boys, it's Pete and Repeat.",
                   "If I had twins with a black girl, this is Tyrone and Tyclone.",
                   "Look at how good my balls are. Look, look. Or because you're Punjabi, look at how good my junk they are."],
                  "Sam_Morril":
                  ["That is male privilege right there, isn't it? I just told a room full of people I got roofied. I did not see one concerned face in here.",
                   "It's hard not to drink man. It's very hard. I work in a place where there's a bar every night. I've been trying to be good. It's very difficult."
                   "That's helpful. Of course you want to get fucked up. Have you tried living? I'm happy and I want to get fucked"],
                  "Sam_Morril_2":
                  ["It's weird that they can do anti-smoking ads, but you can't do pro-cigarette commercials.",
                   "Wouldn't kill you to carry a pack, you know? Marlboro: Just in Case.",
                   "I'm two to three abortions away from Diamond Medallion status, so… I'm losing some of you, but I'm gonna take it a step further here.",
                   "That one looks unfinished. That's what we're going for, so, yeah."],
                  "Trevor_Noah":
                  ["You don't mess with the Russians. Most frightening people in the world. You know how I know this? You know how I've learned? Because I've learned how to use the Russian accent for myself. I've learned how to harness that energy and use it for good.",
                   "I feel safe, like even if there's a monster under the bed, he'd be like, 'Is that a Russian?' There's just something about that Russian accent. Ladies, every single one of you needs to learn the Russian accent."],
                  "Trevor_Noah_2":
                  ["You can't be normal man with a voice like that. You are destined for greatness.",
                   "It just doesn't work. I remember when Nelson Mandela was still alive, and he would tell jokes at press conferences and events, and no one would laugh.",
                   "Everything he did, he'd just be like, 'Ah, knock, knock.' People would be like, 'Let him in. Let him in!' 'No. No, no. You must say, 'Who's there?'' 'He is so right."],
                  "Tom_Segura":
                  ["Oh, shit. Yeah.", "We just need to go get it.",
                   "I'll go get it. You stay here and watch my place.",
                   "That's what's up.",
                   "Can we get a description before we agree to terms? How about a height and weight on old mom?"],
                  "Tom_Segura_2":
                  ["I got such a warm rush through my body. It felt like the inside of my body hugged the outside of my body, you know?",
                   "There's “Hold Open,” and “Close.” And you can watch people walk up and be like, “Mm-mm.” -And you hit that. And then you see it close, and you're like… Sometimes, a second later it opens, and you're like, “Fuck!” You get nervous energy, like you're a kid."]
               }

data = [(comedian, sentence) for comedian, sentences in pandas_dict.items() for sentence in sentences]

gemma_answers = pd.DataFrame(data, columns=['Comedian', 'Sentence'])

gemma_answers.head()

gemma_answers.to_csv('gemma_answers.csv', index=False)

## More Tokens

transcripts = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv')

instruction = "Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments. List of quotes:"

results_dict = {}
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    prompt = f"'''{transcript}'''\n\n{instruction}"
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=250)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(prompt, "").strip()
    results_dict[comedian] = response

for key, value in results_dict.items():
    print(key, value)
    print()

# Sanity Checks with Ted Talks

instruction = "Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments. List of quotes:"

#How do antidepressants work? Ted Talk
ted1= "In the 1950s,  the discovery of two new drugs sparked what would become a multibillion  dollar market for antidepressants. Neither drug was intended  to treat depression at all— in fact, at the time, many doctors  and scientists believed psychotherapy was the only approach  to treating depression. The decades-long journey of discovery that followed revolutionized our understanding  of depression— and raised questions  we hadn’t considered before.One of those first two antidepressant drugs was ipronaizid, which was intended to treat tuberculosis. In a 1952 trial,  it not only treated tuberculosis, it also improved the moods of patients who had previously been diagnosed  with depression. In 1956, a Swiss clinician observed  a similar effect when running a trial for imipramine,  a drug for allergic reactions. Both drugs affected a class  of neurotransmitters called monoamines.The discovery of these  antidepressant drugs gave rise  to the chemical imbalance theory, the idea that depression is caused by having insufficient monoamines in the brain’s synapses. Ipronaizid, imipramine,  and other drugs like them were thought to restore that balance by increasing the availability  of monoamines in the brain.These drugs targeted several different monoamines, each of which acted on a wide range of receptors in the brain. This often meant a lot of side effects, including headaches, grogginess,  and cognitive impairments including difficulty with memory,  thinking, and judgment.Hoping to make the drugs more targeted  and reduce side effects, scientists began studying existing  antidepressants to figure out which specific monoamines  were most associated with improvements in depression. In the 1970s, several different researchers converged on an answer: the most effective antidepressants  all seemed to act on one monoamine called serotonin.This discovery led to the production of fluoxetine, or Prozac, in 1988. It was the first of a new class of drugs called Selective Serotonin  Reuptake Inhibitors, or SSRI’s, which block the reabsorption of serotonin, leaving more available in the brain. Prozac worked well  and had fewer side effects than older, less targeted antidepressants.The makers of Prozac also worked  to market the drug by raising awareness  of the dangers of depression to both the public  and the medical community. More people came to see depression  as a disease caused by mechanisms beyond  an individual’s control, which reduced the culture of blame  and stigmatization surrounding depression, and more people sought help. In the 1990s, the number of people being treated for depression skyrocketed. Psychotherapy and other treatments fell by the wayside, and most people were treated  solely with antidepressant drugs.Since then, we’ve developed a more nuanced view of how to treat depression— and of what causes it. Not everyone with depression responds  to SSRIs like Prozac— some respond better to drugs  that act on other neurotransmitters, or don't respond to medication at all. For many, a combination  of psychotherapy and antidepressant drugs is more effective than either alone. We’re also not sure why antidepressants work the way they do: they change monoamine levels within a few hours of taking the medication, but patients usually don’t feel the benefit until weeks later. And after they stop  taking antidepressants, some patients never experience  depression again, while others relapse.We now recognize that we don’t know what causes depression, or why anti-depressants work. The chemical imbalance theory  is at best an incomplete explanation. It can’t be a coincidence that almost all the antidepressants happen to act on serotonin, but that doesn’t mean serotonin deficiency is the cause of depression. If that sounds odd,  consider a more straightforward example: steroid creams can treat rashes caused by poison ivy— the fact that they work doesn’t mean steroid deficiency was the cause of the rash.We still have a ways to go in terms  of understanding this disease. Fortunately, in the meantime,  we have effective tools to treat it."

input = ted1 + "\n" + instruction
input_ids = tokenizer(input, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=90)
print(tokenizer.decode(outputs[0]))

# How bad data keeps us from good AI
ted2 = "AI could add 16 trillion dollars to the global economy in the next 10 years. This economy is not going to be built by billions of people or millions of factories, but by computers and algorithms. We have already seen amazing benefits of AI in simplifying tasks, bringing efficiencies and improving our lives. However, when it comes to fair and equitable policy decision-making, AI has not lived up to its promise. AI is becoming a gatekeeper to the economy, deciding who gets a job and who gets an access to a loan. AI is only reinforcing and accelerating our bias at speed and scale with societal implications. So, is AI failing us? Are we designing these algorithms to deliver biased and wrong decisions?As a data scientist, I'm here to tell you, it's not the algorithm, but the biased data that's responsible for these decisions. To make AI possible for humanity and society, we need an urgent reset. Instead of algorithms, we need to focus on the data. We're spending time and money to scale AI at the expense of designing and collecting high-quality and contextual data. We need to stop the data, or the biased data that we already have, and focus on three things: data infrastructure, data quality and data literacy.In June of this year, we saw embarrassing bias in the Duke University AI model called PULSE, which enhanced a blurry image into a recognizable photograph of a person. This algorithm incorrectly enhanced a nonwhite image into a Caucasian image. African-American images were underrepresented in the training set, leading to wrong decisions and predictions. Probably this is not the first time you have seen an AI misidentify a Black person's image. Despite an improved AI methodology, the underrepresentation of racial and ethnic populations still left us with biased results.This research is academic, however, not all data biases are academic. Biases have real consequences.Take the 2020 US Census. The census is the foundation for many social and economic policy decisions, therefore the census is required to count 100 percent of the population in the United States. However, with the pandemic and the politics of the citizenship question, undercounting of minorities is a real possibility. I expect significant undercounting of minority groups who are hard to locate, contact, persuade and interview for the census. Undercounting will introduce bias and erode the quality of our data infrastructure.Let's look at undercounts in the 2010 census. 16 million people were omitted in the final counts. This is as large as the total population of Arizona, Arkansas, Oklahoma and Iowa put together for that year. We have also seen about a million kids under the age of five undercounted in the 2010 Census.Now, undercounting of minorities is common in other national censuses, as minorities can be harder to reach, they're mistrustful towards the government or they live in an area under political unrest.For example, the Australian Census in 2016 undercounted Aboriginals and Torres Strait populations by about 17.5 percent. We estimate undercounting in 2020 to be much higher than 2010, and the implications of this bias can be massive.Let's look at the implications of the census data. Census is the most trusted, open and publicly available rich data on population composition and characteristics. While businesses have proprietary information on consumers, the Census Bureau reports definitive, public counts on age, gender, ethnicity, race, employment, family status, as well as geographic distribution, which are the foundation of the population data infrastructure. When minorities are undercounted, AI models supporting public transportation, housing, health care, insurance are likely to overlook the communities that require these services the most.First step to improving results is to make that database representative of age, gender, ethnicity and race per census data. Since census is so important, we have to make every effort to count 100 percent. Investing in this data quality and accuracy is essential to making AI possible, not for only few and privileged, but for everyone in the society.Most AI systems use the data that's already available or collected for some other purposes because it's convenient and cheap. Yet data quality is a discipline that requires commitment — real commitment. This attention to the definition, data collection and measurement of the bias, is not only underappreciated — in the world of speed, scale and convenience, it's often ignored.As part of Nielsen data science team, I went to field visits to collect data, visiting retail stores outside Shanghai and Bangalore. The goal of that visit was to measure retail sales from those stores. We drove miles outside the city, found these small stores — informal, hard to reach. And you may be wondering — why are we interested in these specific stores? We could have selected a store in the city where the electronic data could be easily integrated into a data pipeline — cheap, convenient and easy. Why are we so obsessed with the quality and accuracy of the data from these stores? The answer is simple: because the data from these rural stores matter. According to the International Labour Organization, 40 percent Chinese and 65 percent of Indians live in rural areas. Imagine the bias in decision when 65 percent of consumption in India is excluded in models, meaning the decision will favor the urban over the rural.Without this rural-urban context and signals on livelihood, lifestyle, economy and values, retail brands will make wrong investments on pricing, advertising and marketing. Or the urban bias will lead to wrong rural policy decisions with regards to health and other investments. Wrong decisions are not the problem with the AI algorithm. It's a problem of the data that excludes areas intended to be measured in the first place. The data in the context is a priority, not the algorithms.Let's look at another example. I visited these remote, trailer park homes in Oregon state and New York City apartments to invite these homes to participate in Nielsen panels. Panels are statistically representative samples of homes that we invite to participate in the measurement over a period of time. Our mission to include everybody in the measurement led us to collect data from these Hispanic and African homes who use over-the-air TV reception to an antenna. Per Nielsen data, these homes constitute 15 percent of US households, which is about 45 million people. Commitment and focus on quality means we made every effort to collect information from these 15 percent, hard-to-reach groups.Why does it matter? This is a sizeable group that's very, very important to the marketers, brands, as well as the media companies. Without the data, the marketers and brands and their models would not be able to reach these folks, as well as show ads to these very, very important minority populations. And without the ad revenue, the broadcasters such as Telemundo or Univision, would not be able to deliver free content, including news media, which is so foundational to our democracy.This data is essential for businesses and society. Our once-in-a-lifetime opportunity to reduce human bias in AI starts with the data. Instead of racing to build new algorithms, my mission is to build a better data infrastructure that makes ethical AI possible. I hope you will join me in my mission as well.Thank you."

input = ted2 + "\n" + instruction
input_ids = tokenizer(input, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=90)
print(tokenizer.decode(outputs[0]))

# The fish that walk on land
ted3 = "This pond is the only home  this fish has ever known. But lately, it’s gotten crowded  and food is scarce. Luckily, it has an option many don’t: as a walking catfish, it can dance  its way out of the water and onto bigger and better things. However, it faces many challenges  on its terrestrial journey: it’s now in danger of suffocating,  drying up, suffering physical damage  from rough terrain, and being hunted by land predators.We think of fish  as completely aquatic animals. But the walking catfish is just one  of hundreds of fish species that are actually amphibious, meaning that they possess adaptations  that enable them to survive on land. Fish amphibiousness is a spectrum. At one end are species  like the mosquitofish that’ll only move on land when forced. And at the other end  are species like mudskippers that nonchalantly hop around mudflats  for days at a time. But why do fish make the exodus  from water to land? And how do they cope  with this drastic transition?If temperatures get too high  for the mangrove rivulus in the shallow tropical pools  it inhabits, it’ll flip itself onto a bank  and cool off in the shade. During the dry period, it can survive for two months  out of the water by staying in moist environments. Meanwhile, the eel catfish  makes its onshore voyage to satisfy its hearty craving for beetles. And for others, the terrestrial draw  is more ritualistic. Every year under the cover of night, masses of California grunion flop  their way onto sandy beaches, where females deposit thousands of eggs  into the sand before re-entering the ocean.Underwater, fish breathe with gills, which are feathery organs  packed with blood vessels that absorb dissolved oxygen  from the water. But in the open air, their gills collapse  and are rendered useless, so amphibious fishes need  other ways to breathe. The armored catfish’s stomach  is packed with blood vessels, so it can gulp down air  and breathe through its stomach lining. And lungfish, being related  to the ancestors of all tetrapods, or four-limbed vertebrates,  are equipped with true lungs. They’ll actually drown  if they’re kept underwater too long.Fish have thin, permeable skin  that allows for essential compounds to diffuse into and out of their bodies  while they’re underwater. But this works against them on land  as their bodily moisture diffuses into the air. To dodge dehydration, mudskippers  roll in the mud like puppies. But the lungfish takes the cake: the rivers it inhabits disappear  during dry seasons, so it buries itself in the earth and coats its body in a mucus cocoon. It can survive like this for years  until being resuscitated by the next big rainstorm.Amphibious fishes use powerful fins  to move on land and clever tools to navigate as they go. The Nopoli rock-climbing goby,  no bigger than a few centimeters, scales hundred-meter-tall  Hawaiian waterfalls, inching its way up by alternately  attaching the suction cups on its mouth and pelvic fins. To find water while on land,  the mummichog, like most amphibious fishes,  is on the lookout for reflective surfaces. Other species, like mosquitofish, exercise their inner ear to determine  where they’re oriented on a slope, relying on the probability that they’ll  find water by moving downhill. Our walking catfish, meanwhile, uses the taste buds that coat  its body for navigation. These taste buds are concentrated  in its whiskers, which whip through the air, sensing compounds that signal  the proximity and quality of nearby water— and prey. The walking catfish will shimmy towards  attractive volatile amino acids while steering clear of foul waters  emanating hydrogen sulfide.While amphibious fishes face a multitude  of new challenges upon leaving the water, they’ve evolved ingenious ways  to overcome them. They’re resilient in the face  of droughts and floods and have access to new prey  as well as a plan B if they need to escape competitive,  polluted, or unhealthy environments. While being a “fish out of water”  is generally regarded as a bad thing, for these species,  it offers an undisputed edge."

input = ted3 + "\n" + instruction
input_ids = tokenizer(input, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=90)
print(tokenizer.decode(outputs[0]))

# Asking why?

from collections import defaultdict

results_dict = defaultdict(list)
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    prompt = f"'''{transcript}'''\n\n{instruction}. Explain why they are funny?"
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=350)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(prompt, "").strip()
    results_dict[comedian].append(response)

for comedian, result in results_dict.items():
    print(comedian)
    print(result)
    print()

results_dict = defaultdict(list)
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    prompt = f"'''{transcript}'''\n\n What is this stand-up transcript about?"
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=300)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(prompt, "").strip()
    results_dict[comedian].append(response)

for comedian, result in results_dict.items():
    print(comedian)
    print(result)
    print()
