==> notebooks/03_stand.ipynb <==
# Create Transcript Pandas Dataset

import pandas as pd

from pathlib import Path

aj = Path("/home/ada/humor/data/stand_up_dataset/aj/AJ_TP_text_15.txt").read_text()
aj2 = Path("/home/ada/humor/data/stand_up_dataset/aj2/AJ_TP_text_25.txt").read_text()
aw = Path("/home/ada/humor/data/stand_up_dataset/aw/AW_BC_text_04.txt").read_text()
aw2 = Path("/home/ada/humor/data/stand_up_dataset/aw2/AW_BC_text_25.txt").read_text()
cp = Path("/home/ada/humor/data/stand_up_dataset/cp/CP_OTG_text_18.txt").read_text()
cp2 = Path("/home/ada/humor/data/stand_up_dataset/cp2/CP_OTG_text_07.txt").read_text()
dg = Path("/home/ada/humor/data/stand_up_dataset/dg/DG_W_text_11.txt").read_text()
dg2 = Path("/home/ada/humor/data/stand_up_dataset/dg2/DG_W_text_07.txt").read_text()
hm = Path("/home/ada/humor/data/stand_up_dataset/hm/HM_HK_text_05.txt").read_text()
hm2 = Path("/home/ada/humor/data/stand_up_dataset/hm2/HM_HK_text_22.txt").read_text()
iss = Path("/home/ada/humor/data/stand_up_dataset/is/IS_EM_text_18.txt").read_text()
is2 = Path("/home/ada/humor/data/stand_up_dataset/is2/IS_EM_text_20.txt").read_text()
jg = Path("/home/ada/humor/data/stand_up_dataset/jg/JG_QT_text_03.txt").read_text()
jg2 = Path("/home/ada/humor/data/stand_up_dataset/jg2/JG_QT_text_14.txt").read_text()
ji = Path("/home/ada/humor/data/stand_up_dataset/ji/JL_IHM_text_11.txt").read_text()
ji2 = Path("/home/ada/humor/data/stand_up_dataset/ji2/JL_IHM_text_02.txt").read_text()
jm = Path("/home/ada/humor/data/stand_up_dataset/jm/JB_KGRC_text_12.txt").read_text()
jm2 = Path("/home/ada/humor/data/stand_up_dataset/jm2/JB_KGRC_text_06.txt").read_text()
jy = Path("/home/ada/humor/data/stand_up_dataset/jy/JY_GD_text_09.txt").read_text()
jy2 = Path("/home/ada/humor/data/stand_up_dataset/jy2/JY_GD_text_19.txt").read_text()
lck = Path("/home/ada/humor/data/stand_up_dataset/lck/LCK_SY_text_06.txt").read_text()
lck2 = Path("/home/ada/humor/data/stand_up_dataset/lck2/LCK_SY_text_20.txt").read_text()
nb = Path("/home/ada/humor/data/stand_up_dataset/nb/NB_TGAA_text_09.txt").read_text()
nb2 = Path("/home/ada/humor/data/stand_up_dataset/nb2/NB_TGAA_text_14.txt").read_text()
nb_tk = Path("/home/ada/humor/data/stand_up_dataset/nb_tk/NB_TK_text_20.txt").read_text()
nb_tk2 = Path("/home/ada/humor/data/stand_up_dataset/nb_tk2/NB_TK_text_17.txt").read_text()
rp = Path("/home/ada/humor/data/stand_up_dataset/rp/RP_DP_text_06.txt").read_text()
rp2 = Path("/home/ada/humor/data/stand_up_dataset/rp2/RP_DP_text_21.txt").read_text()
sm = Path("/home/ada/humor/data/stand_up_dataset/sm/SM_IGT_text_04.txt").read_text()
sm2 = Path("/home/ada/humor/data/stand_up_dataset/sm2/SM_IGT_text_13.txt").read_text()
tn = Path("/home/ada/humor/data/stand_up_dataset/tn/TN_AD_text_16.txt").read_text()
tn2 = Path("/home/ada/humor/data/stand_up_dataset/tn2/TN_AD_text_14.txt").read_text()
ts = Path("/home/ada/humor/data/stand_up_dataset/ts/TS_D_text_09.txt").read_text()
ts2 = Path("/home/ada/humor/data/stand_up_dataset/ts2/TS_D_text_15.txt").read_text()

transcripts = [aj, aj2, aw, aw2, cp, cp2, dg, dg2, hm, hm2, iss, is2, jg, jg2, ji, ji2, jm, jm2, jy, jy2, lck, lck2, nb, nb2, nb_tk, nb_tk2, rp, rp2, sm, sm2, tn, tn2, ts, ts2]

len(transcripts)

comedian_names = ["Anthony_Jeselnik", 
                  "Anthony_Jeselnik_2", 
                  "Ali_Wong", 
                  "Ali_Wong_2",
                  "Chelsea_Peretti",
                  "Chelsea_Peretti_2",
                  "Donald_Glover",
                  "Donald_Glover_2",
                  "Hasan_Minhaj",
                  "Hasan_Minhaj_2",
                  "Iliza_Shlesinger",
                  "Iliza_Shlesinger_2",
                  "Jim_Gaffigan",
                  "Jim_Gaffigan_2",
                  "Joe_List",
                  "Joe_List_2",
                  "John_Mulaney",
                  "John_Mulaney_2",
                  "Jimmy_Yang",
                  "Jimmy_Yang_2",
                  "Louis_CK",
                  "Louis_CK_2",
                  "Nate_Bargatze",
                  "Nate_Bargatze_2",
                  "Nate_Bargatze_TK",
                  "Nate_Bargatze_TK_2",
                  "Russell_Peters",
                  "Russell_Peters_2",
                  "Sam_Morril",
                  "Sam_Morril_2",
                  "Trevor_Noah",
                  "Trevor_Noah_2",
                  "Tom_Segura",
                  "Tom_Segura_2"]

len(comedian_names)

stand_up_data = pd.DataFrame({'comedian': comedian_names, 'transcript': transcripts})
print(stand_up_data.head())

stand_up_data.to_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv', index=False)
