==> notebooks/07_prompt_experimentation.ipynb <==
# Experiment with Different Prompts Generated by LLM's

import pandas as pd
import transformers
import inseq
import torch
from pathlib import Path

torch.cuda.is_available()

from huggingface_hub import login
login(token="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")

import pandas as pd

transcripts = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv')

transcripts.head()

def get_quotes(comedian_name, quotes_df):
    quotes_list = quotes_df.loc[quotes_df['comedian'] == comedian_name, 'sentence'].tolist()
    quotes_list = [str(quote) for quote in quotes_list]
    return quotes_list

aj_list = get_quotes("Anthony_Jeselnik", quotes)

aj_list

transcript = transcripts.loc[transcripts['comedian'] == "Anthony_Jeselnik", 'transcript'].iloc[0]

Let's ask ChatGPT for responses

import torch
from transformers import AutoTokenizer, AutoModelForCausalLM

MODEL_ID = "google/gemma-2b-it"

tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
model = AutoModelForCausalLM.from_pretrained(MODEL_ID, device_map="auto")

Let's ask Gemma

prompt = f'''This is a stand up comedy transcript: {transcript}\n 
This is a list of funny quotes: {aj_list}.\n 
What should I ask an AI to get the list of quotes from the transcript?'''

transcript = "When I was a kid, I used to fantasize about getting older, growing up and having money, and buying my mom nice things for her birthday. When I was a kid, we were poor. So poor I remember, just so I could go to my senior prom, just so I could go to my senior prom, I had to sell my U.S. passport on the street. Sold my passport on the street for 300 bucks to get to go to my prom. Of course this was before 9/11 so… my bad, everybody. Weird joke to clap for, but sure. My mom actually should’ve been on one of the planes that crashed on 9/11. I think.I mean, don’t get me wrong, I loved my mother. She was my mom, of course I loved her. We fought a lot. My mom could be very racist. Very racist. And I do not tolerate racism. That’s ignorance and I hate that. When I was a kid, like nine years old, I’d come home after school. Once in a while, I’d bring a friend over to play with me at my house. Once in a while, I’d bring a black friend over. And when I did that, my mom would act weird. She’d pull me aside and say, “Anthony, who’s your new friend? Is he a drug dealer?” And I would say, “Shut up, Mom, that’s racist. Put your money away.”I never get to see my family anymore, really. Most of them are in jail, to be honest. And we never talk, write letters or any of that shit. 'Cause they're all in jail for the exact same thing. My testimony.'"

prompt = f'''This is a stand up comedy transcript: {transcript}\n 
This is a list of funny quotes: {aj_list}.\n 
What should I ask an AI to get the list of quotes from the transcript?'''

input_ids = tokenizer(prompt, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=90)
print(tokenizer.decode(outputs[0]))

1. "To identify the funny quotes from the stand-up comedy transcript, you can ask a language model like me (LLM) the following: "Could you please identify the funny quotes from this stand-up comedy transcript?"

This question prompts the model to analyze the transcript and extract the segments that are intended to be humorous or contain comedic elements. This approach leverages the model's ability to understand context and detect humor based on linguistic cues and typical comedic structures."

prompt = f'''This is an excerpt from a stand-up comedy transcript: {transcript}\n 
            Could you please identify the funny quotes from this stand-up comedy transcript?'''

input_ids = tokenizer(prompt, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=90)
print(tokenizer.decode(outputs[0]))

"What is another prompt I can use?"

prompt = f"Please analyze the following stand-up comedy transcript and identify all instances of humor or funny quotes:{transcript}\n  Please highlight and explain any humorous remarks or quotes found in the transcript."

input_ids = tokenizer(prompt, return_tensors="pt").to("cuda")

outputs = model.generate(**input_ids, max_new_tokens=150)
print(tokenizer.decode(outputs[0]))

transcript = transcripts.loc[transcripts['comedian'] == "Russell_Peters_2", 'transcript'].iloc[0]

Trying Different Prompts
