==> notebooks/10_more_data.ipynb <==
# More Transcripts to Use

Adding more transcripts for experiments

## Imports

import pandas as pd
import transformers
import torch

from huggingface_hub import login
login(token="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")

transcripts = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv')
comedian_name = "Anthony_Jeselnik_3"
start_index = transcripts[transcripts['comedian'] == comedian_name].index[0]
transcripts = transcripts.iloc[start_index:]

## Gemma

import torch
from transformers import AutoTokenizer, AutoModelForCausalLM

MODEL_ID = "google/gemma-2b-it"

tokenizer = AutoTokenizer.from_pretrained(MODEL_ID)
model = AutoModelForCausalLM.from_pretrained(MODEL_ID, device_map="auto")

results = []
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    instruction = f"Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments.{transcript} List of quotes:"
    input_ids = tokenizer(instruction, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=120)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(instruction, "").strip()
    
    results.append({'comedian': comedian, 'sentence': response})

## Random

transcripts = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv')

results = []
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    instruction = f"Extract random sentences from this stand-up comedy transcript. {transcript} List of quotes:"
    input_ids = tokenizer(instruction, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=120)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(instruction, "").strip()
    
    results.append({'comedian': comedian, 'sentence': response})

r = []
seen_quotes = set()
for entry in results:
    comedian = entry['comedian']
    quotes = entry['sentence'].split('\n')  
    
    for quote in quotes:
        if quote and quote not in seen_quotes:
            r.append({'comedian': comedian, 'sentence': quote})
            seen_quotes.add(quote)

quotes_df = pd.DataFrame(r)

quotes_df.to_csv('/home/ada/humor/data/stand_up_dataset/random_gemma_quotes.csv', index=False)

transcripts = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_transcripts.csv')
comedian_name = "Anthony_Jeselnik_3"
start_index = transcripts[transcripts['comedian'] == comedian_name].index[0]
transcripts = transcripts.iloc[start_index:]

## Gemma 2

from transformers import AutoTokenizer, AutoModelForCausalLM
import torch

tokenizer = AutoTokenizer.from_pretrained("google/gemma-2-9b-it")
model = AutoModelForCausalLM.from_pretrained(
    "google/gemma-2-9b-it",
    device_map="auto",
    torch_dtype=torch.bfloat16,
)

results = []
for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    
    instruction = f"Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments.{transcript} List of quotes:"
    input_ids = tokenizer(instruction, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=120)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(instruction, "").strip()
    
    results.append({'comedian': comedian, 'sentence': response})

## Llama

model_id = "meta-llama/Meta-Llama-3-8B-Instruct"

pipeline = transformers.pipeline(
    "text-generation", model=model_id, model_kwargs={"torch_dtype": torch.bfloat16}, device_map="cuda:0"
)

instruction = "Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments. Only list the quotes:"

results_dict = {}

for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript']

    prompt = f"'''{transcript}'''\n\n{instruction}"
    conversation = [{"role": "user", "content": prompt}]
    output = pipeline([conversation], max_new_tokens=120, return_full_text=False)
    
    results_dict[comedian] = output

## Phi

tokenizer = AutoTokenizer.from_pretrained("microsoft/Phi-3-mini-4k-instruct", trust_remote_code=True)
model = AutoModelForCausalLM.from_pretrained("microsoft/Phi-3-mini-4k-instruct", device_map="auto")

results = []
instruction = "Extract the key humorous lines and punchlines for this stand-up comedy transcript. Focus on the quotes highlighting the main comedic moments. List of quotes:"

for index, row in transcripts.iterrows():
    comedian = row['comedian']
    transcript = row['transcript'] 
    prompt = f"'''{transcript}'''\n\n{instruction}"
    
    input_ids = tokenizer(prompt, return_tensors="pt").input_ids.to("cuda")
    outputs = model.generate(input_ids=input_ids, max_new_tokens=120)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    response = generated_text.replace(prompt, "").strip()
    results.append({'comedian': comedian, 'sentence': response})
