==> notebooks/15_spearman.ipynb <==
# Pearson Correlation

import sys
import pandas as pd
sys.path.append("..")
from scipy.stats import pearsonr
from humor.bipartite_metric import bipartite_metric
from humor.vector_similarity_metric import vector_similarity_metric

ground_truth = pd.read_csv('/home/ada/humor/data/stand_up_dataset/standup_data.csv')

gemma2 = pd.read_csv('/home/ada/humor/data/subspace_scores/scores_gemma-2-2b-it.csv')
gemma9 = pd.read_csv('/home/ada/humor/data/subspace_scores/scores_gemma-2-9b-it.csv')
gemma27 = pd.read_csv("/home/ada/humor/data/subspace_scores/scores_gemma-2-27b-it.csv")

gemma2_prompt = pd.read_csv('/home/ada/humor/data/stand_up_dataset/gemma2_2 - gemma2 - gemma2.csv')
gemma9_prompt = pd.read_csv('/home/ada/humor/data/stand_up_dataset/gemma2-9.csv')
gemma27_prompt = pd.read_csv('/home/ada/humor/data/stand_up_dataset/gemma2_27 - gemma2 - gemma2.csv')

2 Billion

print("Subspace\n")
gemma2_score = gemma2["Score"].mean() * 100
print("Gemma 2B Sub: ", gemma2_score)

gemma2_bimetric = bipartite_metric(gemma2_prompt, ground_truth)
print("Gemma 2B Bi: ", gemma2_bimetric.select_dtypes(include='number').mean())

gemma2_vec_metric = vector_similarity_metric(gemma2_prompt, ground_truth)
print("Gemma 2B Vec: ", gemma2_vec_metric.select_dtypes(include='number').mean())

gemma2["Score"] = gemma2["Score"] * 100

# F and V
merged_df4 = pd.merge(gemma2_bimetric, gemma2_vec_metric, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df4['score_df1'], merged_df4['score_df2'])
print("Correlation: ", correlation)

#F and S
gemma2 = gemma2.rename(columns={'Comedian': 'comedian', 'Score': 'score'})
merged_df5 = pd.merge(gemma2_bimetric, gemma2, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df5['score_df1'], merged_df5['score_df2'])
print("Correlation: ", correlation)

#V and S
merged_df6 = pd.merge(gemma2_vec_metric, gemma2, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df6['score_df1'], merged_df6['score_df2'])
print("Correlation: ", correlation)

9 Billion

print("Subspace\n")
gemma9_score = gemma9["Score"].mean() * 100
print("Gemma 9B Sub: ", gemma9_score)

gemma9_bimetric = bipartite_metric(gemma9_prompt, ground_truth)
print("Gemma 9B Bi: ", gemma9_bimetric.select_dtypes(include='number').mean())

gemma9_vec_metric = vector_similarity_metric(gemma9_prompt, ground_truth)
print("Gemma 9B Vec: ", gemma9_vec_metric.select_dtypes(include='number').mean())


gemma9["Score"] = gemma9["Score"] * 100

# F and V
merged_df1 = pd.merge(gemma9_bimetric, gemma9_vec_metric, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df1['score_df1'], merged_df1['score_df2'])
print("Correlation: ", correlation)

#F and S
gemma9 = gemma9.rename(columns={'Comedian': 'comedian', 'Score': 'score'})
merged_df2 = pd.merge(gemma9_bimetric, gemma9, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df2['score_df1'], merged_df2['score_df2'])
print("Correlation: ", correlation)

#V and S
merged_df3 = pd.merge(gemma9_vec_metric, gemma9, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df3['score_df1'], merged_df3['score_df2'])
print("Correlation: ", correlation)

27 billion

gemma27_score = gemma27["Score"].mean() * 100
print("Gemma 27B Sub: ", gemma27_score)

gemma27_bimetric = bipartite_metric(gemma27_prompt, ground_truth)
print("Gemma 27B Bi: ", gemma27_bimetric.select_dtypes(include='number').mean())

gemma27_vec_metric = vector_similarity_metric(gemma27_prompt, ground_truth)
print("Gemma 9B Vec: ", gemma27_vec_metric.select_dtypes(include='number').mean())

gemma27["Score"] = gemma27["Score"] * 100

# F and V
merged_df7 = pd.merge(gemma27_bimetric, gemma27_vec_metric, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df7['score_df1'], merged_df7['score_df2'])
print("Correlation: ", correlation)

#F and S
gemma27 = gemma27.rename(columns={'Comedian': 'comedian', 'Score': 'score'})
merged_df8 = pd.merge(gemma27_bimetric, gemma27, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df8['score_df1'], merged_df8['score_df2'])
print("Correlation: ", correlation)

#V and S
merged_df9 = pd.merge(gemma27_vec_metric, gemma27, on='comedian', suffixes=('_df1', '_df2'))
correlation, _ = pearsonr(merged_df9['score_df1'], merged_df9['score_df2'])
print("Correlation: ", correlation)
